requirements.txt
fastapi==0.111.1
fastapi-cli==0.0.4
Flask==3.0.3
openai==1.37.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
reflex==0.5.4
reflex-hosting-cli==0.1.13
requests==2.32.3
langchain-openai==0.2.3